Move the createUser function to the User object.
